[refac] Moved handshake code out of server.py
